requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.29.0
pgvector==0.3.2
openai==1.51.0
httpx==0.27.0
python-telegram-bot==21.6
pymupdf==1.24.0
ebooklib==0.18
beautifulsoup4==4.12.3
lxml==5.3.0
python-docx==1.1.2
python-pptx==1.0.2
openpyxl==3.1.5
pydantic-settings==2.5.2
pyyaml==6.0.2
python-dotenv==1.0.1
langfuse==3.0.1
